refactor: collapseKeyboardIfFocusOutsideEditText() to Utils.kt
